Mark PowerShell tests as not working on FreeBSD
